parseNum: can now parse numbers with 1 to Long.SIZE (==64) bits. methods for byte, short, int long and char. Parser: s/parsenum.(un)signed/parseNum.(un)signed_int.
statics: renamed char_toInt to char_toNum
added char_printable and a bunch of methods to get the lower or higher
half of a short/int/long. can be used for simple, memory efficient
structures.

StringIterator: for iterating over characters in a String. next() will
suffer frow autoboxing, so use next_ch if possible.

WillNeverThrowException: An unchecked exception with a private
constructor.
and final class.